Use CF Browser Rendering to bypasss website security
This fixes icoi.net and abicdp.org both not getting properly picked up.

Created env vars for browser rendering

For local dev we put `CLOUDFLARE_API_TOKEN` in .dev.vars file and that's already in .gitignore.

For prod we ran `npx wrangler pages secret put CLOUDFLARE_API_TOKEN`.

CLOUDFLARE_ACCOUNT_ID is not a secret so that's why it's in wrangler.jsonc